Created the get_appropriate_direction function
There is now a new constant called GRID_MODE in the Constants class that let you chouse between 3 modes of movements.

When the process_player_input function from the Player script gets the input_direction, it now goes through a _get_appropriate_direction function to get reassigned to a proper value depending on the movement mode.

IMPORTANT: For now, the input_direction conversion is the only thing working. The proper way to handle movement and animation outside of the "GRID_4" movement mode hasn't been implemented.