Implemented grayscale-based transition
Doing so also led to the implementation of an Observer as well as a new script called GDMUtils that could be used to perform various functions.

The implementation of this feature was meant as a first step towards the implementation of map transitions and by extension a functionning world/maps system.