Added map_meta resource (#28)
Added a temporary version of MapMeta without the encounter_table field as the WildEncouterGroup resource hasn't been defined yet. I also made a new enum called MapType for the map_type argument, but I don't really know if it should stay that way. Finally, I decided to not follow the wiki and make the position a Vector2i instead of a Vector2 as I don't see why you'd need floats for the position of the map in the world.